requirements.txt
GitPython>=3.1.40
click>=8.1.7
paramiko>=3.4.0
cryptography>=41.0.7
rich>=13.7.0
colorama>=0.4.6
